refactor(ui): extract summary-page drawing from pdf-export.ts (870→338 lines)
Move combo, tap dance, key override, alt repeat key, and macro
summary-page rendering into pdf-summary-pages.ts (559 lines).
Both files now well within the 800-line Service/Util limit.
Re-exports preserve backward compatibility for all consumers.
